requirements.txt
Flask==3.0.2
adafruit-circuitpython-neopixel==6.3.15
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7 